renamed nSII to k-SII and closes #37
