+++ shapiq/approximator/__init__.py
@@ -1,5 +1,5 @@
 """This module contains the approximators to estimate the Shapley interaction values."""
-from ._base import convert_nsii_into_one_dimension, transforms_sii_to_nsii  # TODO add to tests
+from .k_sii import convert_ksii_into_one_dimension, transforms_sii_to_ksii
 from .permutation.sii import PermutationSamplingSII
 from .permutation.sti import PermutationSamplingSTI
 from .regression import RegressionSII, RegressionFSI, KernelSHAP
@@ -12,6 +12,4 @@ __all__ = [
     "RegressionFSI",
     "RegressionSII",
     "ShapIQ",
-    "transforms_sii_to_nsii",
-    "convert_nsii_into_one_dimension",
 ]
