+++ shapiq/explainer/tree.py
@@ -2,7 +2,7 @@
 for trees and tree ensembles."""
 import numpy as np
 
-from approximator._base import InteractionValues
+from approximator._interaction_values import InteractionValues
 from explainer._base import Explainer
