+++ tests/tests_approximators/test_approximator_permutation_sti.py
@@ -4,7 +4,7 @@ from copy import copy, deepcopy
 import numpy as np
 import pytest
 
-from approximator._base import InteractionValues
+from approximator._interaction_values import InteractionValues
 from approximator.permutation import PermutationSamplingSTI
 from games import DummyGame
