+++ tests/tests_approximators/test_approximator_regression_sv.py
@@ -4,7 +4,7 @@ from copy import deepcopy, copy
 import numpy as np
 import pytest
 
-from approximator._base import InteractionValues
+from approximator._interaction_values import InteractionValues
 from approximator.regression import KernelSHAP
 from games import DummyGame
